Migrated a few more clientside events
 Changes to be committed:
	modified:   fabric-events-lifecycle-v0/src/main/java/net/fabricmc/fabric/api/event/client/LanServerPublishedCallback.java
	modified:   fabric-events-lifecycle-v0/src/main/java/net/fabricmc/fabric/api/event/client/OutOfMemoryCallback.java
	modified:   fabric-events-lifecycle-v0/src/main/java/net/fabricmc/fabric/impl/event/LegacyClientEventInvokers.java
	modified:   fabric-events-lifecycle-v0/src/main/java/net/fabricmc/fabric/impl/event/LegacyEventInvokers.java
	deleted:    fabric-events-lifecycle-v0/src/main/java/net/fabricmc/fabric/mixin/event/lifecycle/MixinMinecraftClient.java
	modified:   fabric-events-lifecycle-v0/src/main/resources/fabric-events-lifecycle-v0.mixins.json
	modified:   fabric-lifecycle-events-v1/src/main/java/net/fabricmc/fabric/api/client/event/lifecycle/v1/ClientLifecycleEvents.java
	renamed:    fabric-events-lifecycle-v0/src/main/java/net/fabricmc/fabric/mixin/event/lifecycle/MixinIntegratedServer.java -> fabric-lifecycle-events-v1/src/main/java/net/fabricmc/fabric/mixin/event/lifecycle/client/MixinIntegratedServer.java
	modified:   fabric-lifecycle-events-v1/src/main/java/net/fabricmc/fabric/mixin/event/lifecycle/client/MixinMinecraftClient.java
	modified:   fabric-lifecycle-events-v1/src/main/resources/fabric-lifecycle-events-v1.mixins.json
