--- fabric-events-lifecycle-v0/src/main/java/net/fabricmc/fabric/mixin/event/lifecycle/MixinIntegratedServer.java
@@ -1,49 +1,0 @@
-/*
- * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package net.fabricmc.fabric.mixin.event.lifecycle;
-
-import org.spongepowered.asm.mixin.Final;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.server.integrated.IntegratedServer;
-import net.minecraft.world.level.LevelInfo;
-
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
-import net.fabricmc.fabric.api.event.client.LanServerPublishedCallback;
-
-@Environment(EnvType.CLIENT)
-@Mixin(IntegratedServer.class)
-public class MixinIntegratedServer {
-	@Shadow
-	@Final
-	private MinecraftClient client;
-
-	@Shadow
-	@Final
-	private LevelInfo levelInfo;
-
-	@Inject(method = "method_6437", at = @At("HEAD"))
-	public void publishToLan(LevelInfo.GameMode gameMode, boolean cheats, CallbackInfoReturnable<String> info) {
-		LanServerPublishedCallback.EVENT.invoker().onServerPublished(this.client, gameMode, cheats, this.levelInfo);
-	}
-}
