+++ fabric-events-lifecycle-v0/src/main/resources/fabric-events-lifecycle-v0.mixins.json
@@ -14,9 +14,7 @@
   ],
   "client": [
     "MixinGameOptions",
-    "MixinIntegratedServer",
-    "MixinItem",
-    "MixinMinecraftClient"
+    "MixinItem"
   ],
   "server": [
     "MixinDedicatedServer"
