+++ fabric-lifecycle-events-v1/src/main/resources/fabric-lifecycle-events-v1.mixins.json
@@ -7,6 +7,7 @@
     "MixinServerWorld"
   ],
   "client": [
+    "client.MixinIntegratedServer",
     "client.MixinMinecraftClient"
   ],
   "injectors": {
